refactor filtering and lists
Since most of filtering is HTML-coupled, most of the filtering logic
has stayed in the HTML context.

List evaluation is much more elegant, and is how things should be.
The logic of deducing the list type is HTML-specific and was kept
in the HTML context. However, the Recursor has been given the job
of evaluating a certain list type as it likes.

In general:
- the Context decides WHAT you are evaluating,
- the Recursor decides HOW you are evaluating it.