start on an AMCSD scraper
Uses a lot of same code from ICSD scraper.

TODO:
Create a generic base class for scrapers, which holds this generic code.
Each scraper should inherit from this, and contain all job-specific information,
such as selectors and scrape logic. Base class should handle non-specific logic.
The goal of this is to make it as easy as possible to write a scraper.